Fixed url handling when url has path.
If url has path we'll decouple path from url and concatenate with
the current path we want to resolve.

If url is https we consider it being returned by directory and we
just returne it.

Also set an expiration date to the order if no expiration date is
returned.

Also ignore tos if 404.

Fixes: #71 #72