Display ids and servers of all accounts.
For all accounts present in the automatoes work directory the list
result will display account id and account server.

We need to figure out what will be the account id because if an
account has the id in two servers, now we have a problem.

This is my last 2022 commit. Happy 2023!!!

Refs: #82